Display better error on missing resources
Instead of a panic because of a nil access, display the reason for the problem. While this would already fail if one uses our make file (go generate fails), one can still manually compile the launcher. As a developer I am currently doing this. So I think it does not hurt to add a little more friendly error message there.